requirements.txt
PyMuPDF>=1.26.4
Pillow>=11.3.0